requirements.txt
python-telegram-bot==20.3
yt-dlp
ffmpeg-python
python-dotenv